requirements.txt
progress
pandas
torch
pillow
numpy
fasttext
tqdm
transformers